back/requirement.txt
flask
flask-cors
flask-sqlalchemy
flask-security-too
flask-bcrypt
redis
celery
flaks-mail